% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(data, init_node)
}
\arguments{
\item{data}{is a data.frame with three columns: v1, v2, and w.}

\item{init_node}{is the node to start from}
}
\value{
Returns a vector with the shortest distance from init_node to all the other nodes in the graph
}
\description{
Calculates the shortest distance from a source node to all other nodes in a graph. The function is built for the
graph that is presented on the wikipedia page for Dijkstra's algorithm
}
\examples{
dijkstra(wiki_graph,1)




}
\references{
\href{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}{Wikipedia page}
}
