% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Title}
\usage{
euclidean(x1, x2)
}
\arguments{
\item{x1}{first number}

\item{x2}{second number}
}
\value{
returns greater common divisor of two value
}
\description{
Title
}
\examples{
euclidean(100, 1000)
}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{wikipedia page}
}
