% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{v1}{price, in US dollars}
  \item{v2}{weight of the diamond, in carats}
  \item{w}{weight of the diamond, in carats}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wiki_graph
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
