% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Dataset for Dijkstra's algorithm presented on the Wikipedia page.}
\format{
A data.frame with 18 rows and 3 columns:
\describe{
  \item{v1}{Number of one vertex in the graph}
  \item{v2}{Number of another vertex in the graph}
  \item{w}{Weight of the edge in the graph connecting two vertices}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the vertices and edges of the graph presented on the wikipedia page on Dijkstra's algorithm.
}
\keyword{datasets}
